+++ app/src/main/java/nextstep/payments/ui/common/PaymentsCard.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.common
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
