+++ app/src/main/java/nextstep/payments/ui/newcard/NewCardActivity.kt
@@ -1,0 +1,17 @@
+package nextstep.payments.ui.newcard
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import nextstep.payments.ui.theme.PaymentsTheme
+
+class NewCardActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            PaymentsTheme {
+                NewCardScreen()
+            }
+        }
+    }
+}
