+++ app/src/main/java/nextstep/payments/ui/newcard/NewCardScreen.kt
@@ -1,4 +1,4 @@
-package nextstep.payments
+package nextstep.payments.ui.newcard
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -18,6 +18,8 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
+import nextstep.payments.ui.common.PaymentCard
+import nextstep.payments.ui.newcard.component.NewCardTopBar
 import nextstep.payments.ui.theme.PaymentsTheme
